refactor code

+++ include/BusySpinWaitStrategy.hpp
@@ -15,10 +15,10 @@ namespace disruptor {
     template<size_t NUMBER_DEPENDENT_SEQUENCES>
     class BusySpinWaitStrategy final : public WaitStrategy<NUMBER_DEPENDENT_SEQUENCES> {
     public:
-        [[nodiscard]] int64_t waitFor(const int64_t sequence,
+        [[nodiscard]] size_t waitFor(const size_t sequence,
                                       SequenceGroupForSingleThread<NUMBER_DEPENDENT_SEQUENCES> &dependent_sequences,
                                       const SequenceBarrier &barrier) override {
-            int64_t availableSequence;
+            size_t availableSequence;
 
             while ((availableSequence = dependent_sequences.get()) < sequence) {
                 barrier.checkAlert();
