+++ include/event.hpp
@@ -3,16 +3,16 @@
 
 namespace disruptor {
     class Event {
-        int64_t value = 0;
+        size_t value = 0;
 
     public:
         Event() = default;
 
-        void setValue(const int64_t value) {
+        void setValue(const size_t value) {
             this->value = value;
         }
 
-        [[nodiscard]] int64_t getValue() const {
+        [[nodiscard]] size_t getValue() const {
             return this->value;
         }
     };
